./scion.sh test works
